fixed history behavior on CTRL+C; fixed empty line print
