+++ inc/control_sequences.h
@@ -6,7 +6,7 @@
 /*   By: mcutura <mcutura@student.42berlin.de>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/18 07:21:52 by mcutura           #+#    #+#             */
-/*   Updated: 2023/06/21 12:22:36 by mcutura          ###   ########.fr       */
+/*   Updated: 2023/06/22 12:48:53 by mcutura          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,12 +27,12 @@
 # define ARROW_LEFT		0x445B1B
 # define DELETE			0x7E335B1B
 
-# define MOVE_LEFT(x)	ft_dprintf(STDOUT_FILENO, "\033[%dD", x)
-# define MOVE_RIGHT(x)	ft_dprintf(STDOUT_FILENO, "\033[%dC", x)
-# define MOVE_COL(x)	ft_dprintf(STDOUT_FILENO, "\033[%dG", x)
-# define CLEAR_LINE		ft_dprintf(STDOUT_FILENO, "\033[2K")
-# define CLEAR_REST		ft_dprintf(STDOUT_FILENO, "\033[0J")
-# define CLEAR_SCREEN	ft_dprintf(STDOUT_FILENO, "\033[2J")
-# define MOVE_HOME		ft_dprintf(STDOUT_FILENO, "\033[H")
+# define MOVE_LEFT(x)	(void)ft_dprintf(STDOUT_FILENO, "\033[%dD", x)
+# define MOVE_RIGHT(x)	(void)ft_dprintf(STDOUT_FILENO, "\033[%dC", x)
+# define MOVE_COL(x)	(void)ft_dprintf(STDOUT_FILENO, "\033[%dG", x)
+# define CLEAR_LINE		(void)ft_dprintf(STDOUT_FILENO, "\033[2K")
+# define CLEAR_REST		(void)ft_dprintf(STDOUT_FILENO, "\033[0J")
+# define CLEAR_SCREEN	(void)ft_dprintf(STDOUT_FILENO, "\033[2J")
+# define MOVE_HOME		(void)ft_dprintf(STDOUT_FILENO, "\033[H")
 
 #endif
